Created API skeleton.
Worked with Carter @grovecj to determine the method functionality for the top-level Pathfinder object, the model objects and the delegate protocols. None of the implementation has been started, let alone completed. Any implementation details in this code are subject to change. Documentation will be added once I determine the best practices for documenting Swift.